Moving generating embeddings task from connection layer to service layer
