+++ src/main/java/org/mule/extension/vectors/internal/connection/model/mistralai/MistralAIModelConnection.java
@@ -6,7 +6,7 @@
  */
 package org.mule.extension.vectors.internal.connection.model.mistralai;
 
-import org.mule.extension.vectors.internal.connection.model.BaseTextModelConnection;
+import org.mule.extension.vectors.internal.connection.model.BaseModelConnection;
 import org.mule.extension.vectors.internal.constant.Constants;
 import org.mule.extension.vectors.internal.error.MuleVectorsErrorType;
 import org.mule.extension.vectors.internal.helper.request.HttpRequestHelper;
@@ -30,7 +30,7 @@ import java.util.concurrent.ExecutionException;
 
 @Alias("mistralAI")
 @DisplayName("Mistral AI")
-public class MistralAIModelConnection implements BaseTextModelConnection {
+public class MistralAIModelConnection implements BaseModelConnection {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MistralAIModelConnection.class);
     private static final String MODELS_ENDPOINT = "https://api.mistral.ai/v1/models";
@@ -50,6 +50,14 @@ public class MistralAIModelConnection implements BaseTextModelConnection {
 
     public String getApiKey() {
         return apiKey;
+    }
+
+    public HttpClient getHttpClient() {
+        return this.httpClient;
+    }
+
+    public long getTimeout() {
+        return this.timeout;
     }
 
     @Override
@@ -80,54 +88,6 @@ public class MistralAIModelConnection implements BaseTextModelConnection {
                         handleErrorResponse(response, "Failed to connect to Mistral AI.");
                     }
                 });
-    }
-
-    @Override
-    public Object generateTextEmbeddings(List<String> inputs, String modelName) {
-        if (inputs == null || inputs.isEmpty()) {
-            throw new IllegalArgumentException("Input list cannot be null or empty");
-        }
-        if (modelName == null || modelName.isEmpty()) {
-            throw new IllegalArgumentException("Model name cannot be null or empty");
-        }
-
-        try {
-            return generateTextEmbeddingsAsync(inputs, modelName).get();
-        } catch (InterruptedException | ExecutionException e) {
-            Thread.currentThread().interrupt();
-            if (e.getCause() instanceof ModuleException) {
-                throw (ModuleException) e.getCause();
-            }
-            throw new ModuleException("Failed to generate embeddings", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
-        }
-    }
-
-    private CompletableFuture<String> generateTextEmbeddingsAsync(List<String> inputs, String modelName) {
-        try {
-            byte[] body = buildEmbeddingsPayload(inputs, modelName);
-            return HttpRequestHelper.executePostRequest(httpClient, EMBEDDINGS_ENDPOINT, buildAuthHeaders(), body, (int) timeout)
-                    .thenApply(this::handleEmbeddingResponse);
-        } catch (JsonProcessingException e) {
-            return CompletableFuture.failedFuture(new ModuleException("Failed to create request body", MuleVectorsErrorType.EMBEDDING_OPERATIONS_FAILURE, e));
-        }
-    }
-
-    private String handleEmbeddingResponse(HttpResponse response) {
-        if (response.getStatusCode() != 200) {
-            return handleErrorResponse(response, "Error generating embeddings");
-        }
-        try {
-            return new String(response.getEntity().getBytes());
-        } catch (IOException e) {
-            throw new ModuleException("Failed to read embedding response", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
-        }
-    }
-
-    private byte[] buildEmbeddingsPayload(List<String> inputs, String modelName) throws JsonProcessingException {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("model", modelName);
-        requestBody.put("input", inputs);
-        return objectMapper.writeValueAsBytes(requestBody);
     }
 
     private Map<String, String> buildAuthHeaders() {
