+++ src/main/java/org/mule/extension/vectors/internal/connection/model/ollama/OllamaModelConnection.java
@@ -8,7 +8,8 @@ package org.mule.extension.vectors.internal.connection.model.ollama;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.mule.extension.vectors.internal.connection.model.BaseTextModelConnection;
+
+import org.mule.extension.vectors.internal.connection.model.BaseModelConnection;
 import org.mule.extension.vectors.internal.constant.Constants;
 import org.mule.extension.vectors.internal.error.MuleVectorsErrorType;
 import org.mule.extension.vectors.internal.helper.request.HttpRequestHelper;
@@ -29,7 +30,7 @@ import java.util.concurrent.ExecutionException;
 
 @Alias("ollama")
 @DisplayName("Ollama")
-public class OllamaModelConnection implements BaseTextModelConnection {
+public class OllamaModelConnection implements BaseModelConnection {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(OllamaModelConnection.class);
 
@@ -53,6 +54,10 @@ public class OllamaModelConnection implements BaseTextModelConnection {
 
     public long getTimeout() {
         return timeout;
+    }
+
+    public HttpClient getHttpClient() {
+        return this.httpClient;
     }
 
     @Override
@@ -83,61 +88,6 @@ public class OllamaModelConnection implements BaseTextModelConnection {
                         handleErrorResponse(response, "Error getting models");
                     }
                 });
-    }
-
-    @Override
-    public Object generateTextEmbeddings(List<String> inputs, String modelName) {
-        if (inputs == null || inputs.isEmpty()) {
-            throw new IllegalArgumentException("Input list cannot be null or empty");
-        }
-        if (inputs.size() != 1) {
-            throw new IllegalArgumentException("Ollama embeddings API accepts only one input at a time");
-        }
-        if (modelName == null || modelName.isEmpty()) {
-            throw new IllegalArgumentException("Model name cannot be null or empty");
-        }
-
-        try {
-            return generateTextEmbeddingsAsync(inputs.get(0), modelName).get();
-        } catch (InterruptedException | ExecutionException e) {
-            Thread.currentThread().interrupt();
-            if (e.getCause() instanceof ModuleException) {
-                throw (ModuleException) e.getCause();
-            }
-            throw new ModuleException("Failed to generate embeddings", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
-        }
-    }
-
-    private CompletableFuture<String> generateTextEmbeddingsAsync(String input, String modelName) {
-        try {
-            byte[] body = buildEmbeddingsPayload(input, modelName);
-            return HttpRequestHelper.executePostRequest(httpClient, baseUrl + EMBEDDINGS_ENDPOINT, buildJsonHeaders(), body, (int) timeout)
-                    .thenApply(this::handleEmbeddingResponse);
-        } catch (JsonProcessingException e) {
-            return CompletableFuture.failedFuture(new ModuleException("Failed to create request body", MuleVectorsErrorType.EMBEDDING_OPERATIONS_FAILURE, e));
-        }
-    }
-
-    private String handleEmbeddingResponse(HttpResponse response) {
-        if (response.getStatusCode() != 200) {
-            return handleErrorResponse(response, "Error generating embeddings");
-        }
-        try {
-            return new String(response.getEntity().getBytes());
-        } catch (IOException e) {
-            throw new ModuleException("Failed to read embedding response", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
-        }
-    }
-
-    private byte[] buildEmbeddingsPayload(String input, String modelName) throws JsonProcessingException {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("model", modelName);
-        requestBody.put("prompt", input); // Ollama takes one input at a time
-        return objectMapper.writeValueAsBytes(requestBody);
-    }
-
-    private Map<String, String> buildJsonHeaders() {
-        return Map.of("Content-Type", "application/json");
     }
 
     private String handleErrorResponse(HttpResponse response, String message) {
