+++ src/main/java/org/mule/extension/vectors/internal/model/azureopenai/AzureOpenAIService.java
@@ -7,24 +7,120 @@ import dev.langchain4j.model.output.TokenUsage;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.mule.extension.vectors.internal.connection.model.azureopenai.AzureOpenAIModelConnection;
+import org.mule.extension.vectors.internal.error.MuleVectorsErrorType;
 import org.mule.extension.vectors.internal.helper.parameter.EmbeddingModelParameters;
+import org.mule.extension.vectors.internal.helper.request.HttpRequestHelper;
+import org.mule.extension.vectors.internal.model.nomic.NomicService;
 import org.mule.extension.vectors.internal.service.embedding.EmbeddingService;
+import org.mule.runtime.extension.api.exception.ModuleException;
+import org.mule.runtime.http.api.domain.message.response.HttpResponse;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class AzureOpenAIService implements EmbeddingService {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(AzureOpenAIService.class);
   private AzureOpenAIModelConnection azureOpenAIModelConnection;
   private EmbeddingModelParameters embeddingModelParameters;
   private Integer dimensions;
   private static final int BATCH_SIZE = 16;
+  private final ObjectMapper objectMapper = new ObjectMapper();
 
   public AzureOpenAIService(AzureOpenAIModelConnection azureOpenAIModelConnection, EmbeddingModelParameters embeddingModelParameters, Integer dimensions) {
     this.azureOpenAIModelConnection = azureOpenAIModelConnection;
     this.embeddingModelParameters = embeddingModelParameters;
     this.dimensions = dimensions;
+  }
+
+  public Object generateTextEmbeddings(List<String> inputs, String deploymentName) {
+      if (inputs == null || inputs.isEmpty()) {
+          throw new IllegalArgumentException("Input list cannot be null or empty");
+      }
+      if (deploymentName == null || deploymentName.isEmpty()) {
+          throw new IllegalArgumentException("Deployment name cannot be null or empty");
+      }
+      try {
+          return generateTextEmbeddingsAsync(inputs, deploymentName).get();
+      } catch (InterruptedException | ExecutionException e) {
+          Thread.currentThread().interrupt();
+          if (e.getCause() instanceof ModuleException) {
+              throw (ModuleException) e.getCause();
+          }
+          throw new ModuleException("Failed to generate embeddings", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
+      }
+  }
+
+  private CompletableFuture<String> generateTextEmbeddingsAsync(List<String> inputs, String deploymentName) {
+      String url = buildUrlForDeployment(deploymentName);
+      try {
+          byte[] body = buildTextEmbeddingPayload(inputs);
+          return HttpRequestHelper.executePostRequest(this.azureOpenAIModelConnection.getHttpClient(), url, buildHeaders(), body, (int) this.azureOpenAIModelConnection.getTimeout())
+                  .thenApply(this::handleEmbeddingResponse);
+      } catch (JsonProcessingException e) {
+          return CompletableFuture.failedFuture(new ModuleException("Failed to create request body", MuleVectorsErrorType.EMBEDDING_OPERATIONS_FAILURE, e));
+      }
+  }
+
+  private String handleEmbeddingResponse(HttpResponse response) {
+      if (response.getStatusCode() != 200) {
+          return handleErrorResponse(response, "Error generating embeddings");
+      }
+      try {
+          return new String(response.getEntity().getBytes(), StandardCharsets.UTF_8);
+      } catch (IOException e) {
+          throw new ModuleException("Failed to read embedding response", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
+      }
+  }
+
+  private String buildUrlForDeployment(String deploymentName) {
+      try {
+          String encodedDeployment = URLEncoder.encode(deploymentName, StandardCharsets.UTF_8.name());
+          String encodedApiVersion = URLEncoder.encode(this.azureOpenAIModelConnection.getApiVersion(), StandardCharsets.UTF_8.name());
+          return String.format("%s/openai/deployments/%s/embeddings?api-version=%s", this.azureOpenAIModelConnection.getEndpoint(), encodedDeployment, encodedApiVersion);
+      } catch (UnsupportedEncodingException e) {
+          throw new ModuleException("Failed to encode URL parameters", MuleVectorsErrorType.EMBEDDING_OPERATIONS_FAILURE, e);
+      }
+  }
+
+  private byte[] buildTextEmbeddingPayload(List<String> inputs) throws JsonProcessingException {
+      Map<String, Object> requestBody = new HashMap<>();
+      requestBody.put("input", inputs);
+      return objectMapper.writeValueAsBytes(requestBody);
+  }
+  
+  private Map<String, String> buildHeaders() {
+      Map<String, String> headers = new HashMap<>();
+      headers.put("api-key", this.azureOpenAIModelConnection.getApiKey());
+      headers.put("Content-Type", "application/json");
+      return headers;
+  }
+
+  private String handleErrorResponse(HttpResponse response, String message) {
+      try {
+          String errorBody = new String(response.getEntity().getBytes(), StandardCharsets.UTF_8);
+          String errorMsg = String.format("%s. Azure OpenAI API error (HTTP %d): %s",
+                  message, response.getStatusCode(), errorBody);
+          LOGGER.error(errorMsg);
+          throw new ModuleException(errorMsg, MuleVectorsErrorType.AI_SERVICES_FAILURE);
+      } catch (IOException e) {
+          throw new ModuleException("Failed to read error response body", MuleVectorsErrorType.AI_SERVICES_FAILURE, e);
+      }
   }
 
   @Override
@@ -39,7 +135,7 @@ public class AzureOpenAIService implements EmbeddingService {
         List<String> batch = texts.subList(x, Math.min(x + BATCH_SIZE, texts.size()));
 
         // Generate embeddings for current batch
-        String response = (String) this.azureOpenAIModelConnection.generateTextEmbeddings(batch, embeddingModelParameters.getEmbeddingModelName());
+        String response = (String) generateTextEmbeddings(batch, embeddingModelParameters.getEmbeddingModelName());
         JSONObject jsonResponse = new JSONObject(response);
 
         // Accumulate token usage
